[MCUX-58516][demo_apps][lin_stack] Fix issue that LED 1 never lights
Root cause: On KW45B41Z83, the TPM CnV register is __IO uint32_t, but
MOD is set to 0xFFFF so the counter only cycles 0–65535. In the ISR, the
output-compare is:
CnV = counterVal + DEMO_MODULO_VALUE;  // up to 71534 as a 32-bit value
Once counterVal > 59536 (≈ 5 ms after startup), the sum exceeds 65535
and the 32-bit CnV is written with a value the counter can never reach.
The output compare stops firing permanently after ~11 firings, and before
interruptCount ever reaches 100, g_motorTickCount stays at 10, master
always sees a value below MOTOR1_MIN_VALUE (50) and keeps sending
INCREASE, so g_motorSelectionCmd never becomes DECREASE, and LED1
never lights.

Also update the readme to describe correct behavior, and move the common
part from board specifc read me to the common example readme.

+++ demo_apps/lin_stack/slave/lin_stack_slave.c
@@ -77,7 +77,7 @@ void DEMO_TPM_IRQHandler(void)
 
         TPM_ClearStatusFlags(DEMO_TPM_BASEADDR, DEMO_TPM_CH_OUT_FLG);
         counterVal                                             = (uint16_t)(DEMO_TPM_BASEADDR->CNT);
-        (DEMO_TPM_BASEADDR->CONTROLS[DEMO_TPM_CH_OUT_NUM].CnV) = counterVal + DEMO_MODULO_VALUE;
+        (DEMO_TPM_BASEADDR->CONTROLS[DEMO_TPM_CH_OUT_NUM].CnV) = (uint16_t)(counterVal + DEMO_MODULO_VALUE);
 
         /* Timer Interrupt Handler. */
         lin_lld_timeout_service(LI0);
